feat: hide lifetime membership button
