+++ src/utils/cio/subscriber-attributes.ts
@@ -1,0 +1,3 @@
+const INTERESTED_IN_LIFETIME_SINCE = 'interested_in_lifetime_since'
+
+export {INTERESTED_IN_LIFETIME_SINCE}
